+++ project-euler.d/source/solutions/_12.d
@@ -6,8 +6,8 @@ import euler.common.numbers;
 
 mixin solution!(12,
 {
-	int number = 0;
-	int term   = 0;
+	int  number = 0;
+	long term   = 0;
 	while (true) {
 		number++;
 		term += number;
