+++ project-euler.d/source/solutions/_19.d
@@ -6,7 +6,7 @@ import std.datetime;
 
 mixin solution!(19,
 {
-    int sundays_first = 0;
+    long sundays_first = 0;
     for (int year = 1901; year <= 2000; year++) {
         for (int month = 1; month <= 12; month++) {
             if (Date(year, month, 1).dayOfWeek == DayOfWeek.sun) {
